Bug fix on File Delete

+++ src/Marca/FileBundle/Controller/FileController.php
@@ -213,7 +213,9 @@ class FileController extends Controller
             }
 
             $em->remove($entity);
+            if ($doc) {
             $em->remove($doc);
+            }
             $em->flush();
         }
